+++ examples/simple/main.go
@@ -7,17 +7,17 @@ import (
 )
 
 var (
-	stateOptions = []*tforms.SelectableOption[string]{
+	stateOptions = []*tforms.SelectableOption{
 		{Label: "Alaska", Value: "AK"},
 		{Label: "Georgia", Value: "GA"},
 		{Label: "Oregon", Value: "OG"},
 	}
-	sizesOptions = []*tforms.SelectableOption[float64]{
+	sizesOptions = []*tforms.SelectableOption{
 		{Label: "SM", Value: 0.43},
 		{Label: "MD", Value: 0.24},
 		{Label: "LG", Value: 0.83},
 	}
-	departmentsOptions = []*tforms.SelectableOption[string]{
+	departmentsOptions = []*tforms.SelectableOption{
 		{Label: "Workshop", Value: "workshop"},
 		{Label: "Inventory", Value: "inventory"},
 		{Label: "Sales", Value: "sales"},
@@ -36,6 +36,19 @@ func main() {
 	}
 }
 
+/*
+Use to add validation tags
+*/
+type Employee struct {
+	Email       string    `form:"email" validate:"required,email"`
+	Address     string    `form:"address" validate:"required"`
+	Salary      float64   `form:"salary" validate:"required,min=1000,max=10000"`
+	State       string    `form:"state" validate:"required"`
+	Sizes       []float64 `form:"sizes" validate:"required"`
+	Departments []string  `form:"departments" validate:"required,dive,alphanum"`
+	Description string    `form:"description"`
+}
+
 func handlePostAction(w http.ResponseWriter, r *http.Request) {
 	err := r.ParseForm()
 	if err != nil {
@@ -44,7 +57,15 @@ func handlePostAction(w http.ResponseWriter, r *http.Request) {
 	}
 	form := MakeForm()
 	form.SetValue(r.Form)
-	form.Validate()
+
+	// NOTE: this library is not for validating. It's only for rendering.
+	// 		 For validating use:
+	//		 - github.com/go-playground/validator/v10
+	//		 - github.com/go-ozzo/ozzo-validation
+	// 		 or other. Form takes map[string]string as an error and match errors for fields.
+	form.SetErrors(map[string]string{
+		"email": "invalid email",
+	})
 	IndexPage(form, true).Render(r.Context(), w)
 }
 
@@ -58,19 +79,22 @@ func MakeForm() tforms.IForm {
 		"/post/action",
 		false,
 		tforms.NewHiddenField("_csrf", "csrf-token-string", true),
-		tforms.NewInputField[string]("email", tforms.TextInputEmail, true).
+		tforms.NewInputField("email", tforms.TextInputEmail, true).
 			SetPlaceholder("example@mail.com").
 			SetLabel("Employee email"),
-		tforms.NewInputField[string]("address", tforms.TextInputText, true).
+		tforms.NewInputField("address", tforms.TextInputText, true).
 			SetPlaceholder("st.84, New York, USA").
 			SetLabel("Employee address"),
-		tforms.NewSelectField[string]("state", stateOptions, false, true).
+		tforms.NewInputField("salary", tforms.TextInputNumber, true).
+			SetPlaceholder("$10000").
+			SetLabel("Salary"),
+		tforms.NewSelectField("state", stateOptions, false, true).
 			SetPlaceholder("Select state...").
 			SetLabel("State where employee leaves"),
-		tforms.NewSelectField[float64]("sizes", sizesOptions, false, true).
+		tforms.NewSelectField("sizes", sizesOptions, false, true).
 			SetPlaceholder("Select sizes...").
 			SetLabel("Employee size for work clothes"),
-		tforms.NewSelectableField[string]("departments", tforms.InputTypeCheckbox, departmentsOptions, true, true).
+		tforms.NewSelectableField("departments", tforms.InputTypeCheckbox, departmentsOptions, true, true).
 			SetLabel("Departments"),
 		tforms.NewTextArea("description", true).
 			SetLabel("Employee description"),