+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.2.1 (2025-03-14)
+
+### Fix
+
+- **types**: remove generics, fix setvalue for checkbox
+
+## v0.2.0 (2025-03-08)
+
+### Feat
+
+- remove validation logic + provide public interface for setting errors
+
 ## v0.1.0 (2025-03-08)
 
 ### Feat